Checkpointing: Add testing for emitting state in Destination Standard Test; Migrate BQ to support checkpointing (#3546)
* test checkpointing

* code review feedback

* add checks for state output in standard test

* migrate bigquery to output state

* code review feedback

* cleaned
