+++ airbyte-integrations/bases/base-java/src/main/java/io/airbyte/integrations/destination/buffered_stream_consumer/BufferedStreamConsumer.java
@@ -35,7 +35,6 @@ import io.airbyte.integrations.base.FailureTrackingAirbyteMessageConsumer;
 import io.airbyte.protocol.models.AirbyteMessage;
 import io.airbyte.protocol.models.AirbyteMessage.Type;
 import io.airbyte.protocol.models.AirbyteRecordMessage;
-import io.airbyte.protocol.models.AirbyteStateMessage;
 import io.airbyte.protocol.models.ConfiguredAirbyteCatalog;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -81,13 +80,13 @@ public class BufferedStreamConsumer extends FailureTrackingAirbyteMessageConsume
   private final ConfiguredAirbyteCatalog catalog;
   private final CheckedFunction<String, Boolean, Exception> isValidRecord;
   private final Map<AirbyteStreamNameNamespacePair, Long> pairToIgnoredRecordCount;
-  private final Consumer<AirbyteStateMessage> checkpointConsumer;
+  private final Consumer<AirbyteMessage> outputRecordCollector;
 
   private boolean hasStarted;
   private boolean hasClosed;
 
-  private AirbyteStateMessage lastCommittedState;
-  private AirbyteStateMessage pendingState;
+  private AirbyteMessage lastCommittedState;
+  private AirbyteMessage pendingState;
 
   public BufferedStreamConsumer(Consumer<AirbyteMessage> outputRecordCollector,
                                 VoidCallable onStart,
@@ -95,7 +94,7 @@ public class BufferedStreamConsumer extends FailureTrackingAirbyteMessageConsume
                                 CheckedConsumer<Boolean, Exception> onClose,
                                 ConfiguredAirbyteCatalog catalog,
                                 CheckedFunction<String, Boolean, Exception> isValidRecord) {
-    this.checkpointConsumer = (message) -> outputRecordCollector.accept(new AirbyteMessage().withType(Type.STATE).withState(message));
+    this.outputRecordCollector = outputRecordCollector;
     this.hasStarted = false;
     this.hasClosed = false;
     this.onStart = onStart;
@@ -145,7 +144,7 @@ public class BufferedStreamConsumer extends FailureTrackingAirbyteMessageConsume
         flushQueueToDestination();
       }
     } else if (message.getType() == Type.STATE) {
-      pendingState = message.getState();
+      pendingState = message;
     } else {
       LOGGER.warn("Unexpected message: " + message.getType());
     }
@@ -196,7 +195,7 @@ public class BufferedStreamConsumer extends FailureTrackingAirbyteMessageConsume
       // enable checkpointing, we will need to get feedback from onClose on whether any data was persisted
       // or not. If it was then, the state message will be emitted.
       if (!hasFailed && lastCommittedState != null) {
-        checkpointConsumer.accept(lastCommittedState);
+        outputRecordCollector.accept(lastCommittedState);
       }
     } catch (Exception e) {
       LOGGER.error("on close failed.");
